[dg] RFC: Rename scaffoldalbe to folable and rename the framework fold
