+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/shell-script-component/with-scaffolder.py
@@ -13,9 +13,9 @@ from dagster_components import (
     ScaffoldRequest,
     scaffold_component_yaml,
 )
+from dagster_components.fold.decorator import foldable
 from dagster_components.resolved.core_models import AssetSpecSequenceField
 from dagster_components.resolved.model import ResolvableModel, ResolvedFrom
-from dagster_components.scaffoldable.decorator import scaffoldable
 
 import dagster as dg
 
@@ -48,7 +48,7 @@ class ShellCommandModel(ResolvableModel):
 
 
 # highlight-start
-@scaffoldable(scaffolder=ShellCommandScaffolder)
+@foldable(scaffolder=ShellCommandScaffolder)
 # highlight-end
 @dataclass
 class ShellCommand(Component, ResolvedFrom[ShellCommandModel]):
