+++ python_modules/libraries/dagster-components/dagster_components/components/dbt_project/component.py
@@ -15,6 +15,7 @@ from dagster_dbt import (
 
 from dagster_components import Component, ComponentLoadContext
 from dagster_components.components.dbt_project.scaffolder import DbtProjectComponentScaffolder
+from dagster_components.fold.decorator import foldable
 from dagster_components.resolved.core_models import (
     AssetAttributesModel,
     AssetPostProcessor,
@@ -25,7 +26,6 @@ from dagster_components.resolved.core_models import (
 )
 from dagster_components.resolved.metadata import ResolvableFieldInfo
 from dagster_components.resolved.model import FieldResolver, ResolvableModel, ResolvedFrom
-from dagster_components.scaffoldable.decorator import scaffoldable
 from dagster_components.utils import TranslatorResolvingInfo, get_wrapped_translator_class
 
 
@@ -56,7 +56,7 @@ def resolve_dbt(context: ResolutionContext, dbt: DbtCliResource) -> DbtCliResour
     return DbtCliResource(**context.resolve_value(dbt.model_dump()))
 
 
-@scaffoldable(scaffolder=DbtProjectComponentScaffolder)
+@foldable(scaffolder=DbtProjectComponentScaffolder)
 @dataclass
 class DbtProjectComponent(Component, ResolvedFrom[DbtProjectModel]):
     """Expose a DBT project to Dagster as a set of assets."""
