+++ python_modules/libraries/dagster-components/dagster_components/components/definitions_component/component.py
@@ -12,11 +12,11 @@ from dagster_components import Component, ComponentLoadContext
 from dagster_components.components.definitions_component.scaffolder import (
     DefinitionsComponentScaffolder,
 )
+from dagster_components.fold.decorator import foldable
 from dagster_components.resolved.model import ResolvableModel
-from dagster_components.scaffoldable.decorator import scaffoldable
 
 
-@scaffoldable(scaffolder=DefinitionsComponentScaffolder)
+@foldable(scaffolder=DefinitionsComponentScaffolder)
 class DefinitionsComponent(Component, ResolvableModel):
     """Wraps an arbitrary set of Dagster definitions."""
