+++ python_modules/libraries/dagster-components/dagster_components/components/definitions_component/scaffolder.py
@@ -5,8 +5,8 @@ from dagster._utils import pushd
 from pydantic import BaseModel
 
 from dagster_components.core.component_scaffolder import ScaffoldRequest
+from dagster_components.fold.scaffolder import Scaffolder
 from dagster_components.scaffold import scaffold_component_yaml
-from dagster_components.scaffoldable.scaffolder import Scaffolder
 
 
 class DefinitionsScaffoldParams(BaseModel):
