+++ python_modules/libraries/dagster-components/dagster_components/components/shim_components/asset.py
@@ -4,7 +4,7 @@ from dagster._utils import pushd
 
 from dagster_components import Scaffolder, ScaffoldRequest
 from dagster_components.components.shim_components.base import ShimComponent
-from dagster_components.scaffoldable.decorator import scaffoldable
+from dagster_components.fold.decorator import foldable
 
 
 class AssetScaffolder(Scaffolder):
@@ -18,6 +18,6 @@ class AssetScaffolder(Scaffolder):
 """)
 
 
-@scaffoldable(AssetScaffolder)
+@foldable(AssetScaffolder)
 class RawAssetComponent(ShimComponent):
     """Asset definition component."""
