+++ python_modules/libraries/dagster-components/dagster_components/core/component.py
@@ -20,10 +20,10 @@ from typing_extensions import Self
 
 from dagster_components.core.component_key import ComponentKey
 from dagster_components.core.component_scaffolder import DefaultComponentScaffolder
+from dagster_components.fold.decorator import foldable, get_scaffolder
+from dagster_components.fold.scaffolder import ScaffolderUnavailableReason
 from dagster_components.resolved.context import ResolutionContext
 from dagster_components.resolved.model import ResolvableModel, ResolvedFrom, resolve_model
-from dagster_components.scaffoldable.decorator import get_scaffolder, scaffoldable
-from dagster_components.scaffoldable.scaffolder import ScaffolderUnavailableReason
 from dagster_components.utils import format_error_message
 
 
@@ -39,7 +39,7 @@ class ComponentDeclNode(ABC):
     def get_source_position_tree(self) -> Optional[SourcePositionTree]: ...
 
 
-@scaffoldable(scaffolder=DefaultComponentScaffolder)
+@foldable(scaffolder=DefaultComponentScaffolder)
 class Component(ABC):
     @classmethod
     def get_schema(cls) -> Optional[type["ResolvableModel"]]:
