+++ python_modules/libraries/dagster-components/dagster_components/core/component_scaffolder.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from dagster_components.scaffoldable.scaffolder import Scaffolder, ScaffoldRequest
+from dagster_components.fold.scaffolder import Scaffolder, ScaffoldRequest
 
 
 class DefaultComponentScaffolder(Scaffolder):
