+++ python_modules/libraries/dagster-components/dagster_components_tests/scaffoldable_tests/test_decorator.py
@@ -1,11 +1,7 @@
 import pytest
 from dagster._check.functions import CheckError
-from dagster_components.scaffoldable.decorator import (
-    get_scaffolder,
-    is_scaffoldable_class,
-    scaffoldable,
-)
-from dagster_components.scaffoldable.scaffolder import Scaffolder
+from dagster_components.fold.decorator import foldable, get_scaffolder, is_foldable
+from dagster_components.fold.scaffolder import Scaffolder
 
 
 # Example usage:
@@ -15,7 +11,7 @@ def test_basic_usage() -> None:
         pass
 
     # Example decorated class
-    @scaffoldable(MyScaffolder)
+    @foldable(MyScaffolder)
     class MyClass:
         pass
 
@@ -24,8 +20,8 @@ def test_basic_usage() -> None:
         pass
 
     # Test the functions
-    assert is_scaffoldable_class(MyClass) is True
-    assert is_scaffoldable_class(RegularClass) is False
+    assert is_foldable(MyClass) is True
+    assert is_foldable(RegularClass) is False
     assert isinstance(get_scaffolder(MyClass), MyScaffolder)
     with pytest.raises(CheckError):
         get_scaffolder(RegularClass)
@@ -36,14 +32,14 @@ def test_inheritance() -> None:
 
     class ScaffoldableTwo(Scaffolder): ...
 
-    @scaffoldable(ScaffoldableOne)
+    @foldable(ScaffoldableOne)
     class ClassOne: ...
 
-    @scaffoldable(ScaffoldableTwo)
+    @foldable(ScaffoldableTwo)
     class ClassTwo(ClassOne): ...
 
-    assert is_scaffoldable_class(ClassOne) is True
+    assert is_foldable(ClassOne) is True
     assert isinstance(get_scaffolder(ClassOne), ScaffoldableOne)
 
-    assert is_scaffoldable_class(ClassTwo) is True
+    assert is_foldable(ClassTwo) is True
     assert isinstance(get_scaffolder(ClassTwo), ScaffoldableTwo)
